Rework fin types to just "Flat" and "Standard" airfoils
Flat plate remains the same, while Standard becomes a customizable airfoil that can perform both as a symmetrical and cambered airfoil depending on the camber setting. Increasing the camber slightly increases the angle at which the airfoil stalls in positive AoA but strongly decreases it in negative AoA. Need to make some of these values less hard coded but this works for now.